% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/households.R
\docType{data}
\name{households}
\alias{households}
\title{households: Household survey data}
\format{
A tibble with 116 rows and 89 variables
\describe{
\item{id}{identification number}
\item{community}{the communities surveyed, options including \code{1} kg: Korle Gonno and \code{2} abuja: Abuja}
\item{housing_type}{housing type, options includin \code{1} block_unit: unit in a row of apartments made of cement blocks, \code{2} wood_unit: unit in a row of apartments made of wood, \code{3} house, \code{4} compound_house: single-story L- or C-shaped house with a multiple units around a shared courtyard, \code{5} multistory_apt: multi-story apartment building, \code{6} wood_shack: wooden shack, \code{7} no_structure, and \code{8} other}
\item{respondent_relationship_to_hh}{respondent's relationship to the household head (respondent identified), options including \code{1} self, \code{2} child, \code{3} spouse, and \code{4} other_relative}
\item{gender}{gender (self-identified) of respondent, options including \code{1} female and \code{2} male}
\item{tenure}{tenure status, options including \code{1} rented: renter, \code{2} owned: homeowner, or \code{3} no_payment: living without payment)}
\item{years_in_community}{number of years respondent has lived in community}
\item{adult_count}{number of adults in household including respondent. Household is described as those "eating from the same pot"}
\item{child_count}{number of children under 18 in household. Household is described as those "eating from the same pot"}
\item{rooms_in_hh}{number of rooms used for sleeping. Household is described as those "eating from the same pot"}
\item{business_ownership}{household or respondent owns a business, options including \code{1} respondent-owned and \code{2} household-owned}
\item{business_location}{location type of the business, options including \code{1} home_based, \code{2} outside_home: fixed location outside home, or \code{3} mobile: mobile location.}
\item{business_category}{type of business, options including \code{1} food, \code{2} shop, \code{3} salon, \code{4} vented_water, \code{5} tailoring, and \code{6} other_services.}
\item{business_water_use}{respondent's business uses water beyond typical needs of household (true or false)}
\item{business_water_source}{primary source of water for business use (packaged water, piped to home, piped to neighbor's home, piped to compound, commercial or public tap, borehole, dug well, spring water, delivered water)}
\item{primary_dw_source}{primary source of drinking water (packaged water, piped to home, piped to neighbor's home, piped to compound, commercial or public tap, borehole, dug well, spring water, delivered water)}
\item{dw_reason_convenience}{respondent names convenience as a reason for using the drinking water source (true or false)}
\item{dw_reason_affordable}{respondent names affordability as a reason for using the drinking water source (true or false)}
\item{dw_reason_available}{respondent names availability as a reason for using the drinking water source (true or false)}
\item{dw_reason_cold}{respondent names temperature (cold water) as a reason for using the drinking water source (true or false)}
\item{dw_reason_clean}{respondent names cleanliness as a reason for using the drinking water source (true or false)}
\item{dw_reason_taste}{respondent names taste as a reason for using the drinking water source (true or false)}
\item{dw_reason_habit_or_cultural_norm}{respondent names habit or cultural norm as a reason for using the drinking water source (true or false)}
\item{dw_reason_trustworthy}{respondent names trustworthiness as a reason for using the drinking water source (true or false)}
\item{dw_reason_health}{respondent names health as a reason for using the drinking water source (true or false)}
\item{dw_reason_other}{respondent names another reason for using the drinking water source (true or false)}
\item{package_type_preference}{respondent typically purchases individual, options including \code{1}individual: sachets/packets/bottles, \code{2} bag: multipacks of these, or \code{3} both}
\item{package_size_reason_storage_space}{respondent names storage space in the home as a reason for purchasing the preferred package type (true or false)}
\item{package_size_reason_cost_effective}{respondent names cost effectiveness as a reason for purchasing the preferred package type (true or false)}
\item{package_size_reason_temperature}{respondent names temperature at the time of purchase as a reason for purchasing the preferred package type (true or false)}
\item{package_size_reason_available_money}{respondent names availability of money as a reason for purchasing the preferred package type (true or false)}
\item{package_size_reason_convenience}{respondent names convenience as a reason for purchasing the preferred package type (true or false)}
\item{package_size_reason_size}{respondent names the size needed for the respondent or household as a reason for purchasing the preferred package type (true or false)}
\item{package_size_reason_avoid_wasting_water}{respondent names avoiding wasting water by purchasing only when needed as a reason for purchasing the preferred package type (true or false)}
\item{dw_treatment}{treatment methods of water before drinking, options including \code{1}no_treatment, \code{2} boil, \code{3}boil;settle, \code{4} filter, and \code{5}settle}
\item{primary_water_source}{primary water source for non-drinking water, options including \code{1}packaged water, \code{2}piped_to_home, \code{3}piped to neighbor's home, \code{4}piped to compound, \code{5}commercial or public tap, \code{6}borehole, \code{7}dug well, \code{8}spring water, and \code{9} delivered water)}
\item{primary_source_reason_proximity}{respondent names proximity to home as a reason for using the primary source of non-drinking water (true or false)}
\item{primary_source_reason_convenience}{respondent names convenience as a reason for using the primary source of non-drinking water (true or false)}
\item{primary_source_reason_affordable}{respondent names affordability as a reason for using the primary source of non-drinking water (true or false)}
\item{primary_source_reason_availability}{respondent names availability as a reason for using the primary source of non-drinking water (true or false)}
\item{primary_source_reason_cleanliness}{respondent names cleanliness as a reason for using the primary source of non-drinking water (true or false)}
\item{primary_source_reason_other}{respondent names another reason for using the primary source of non-drinking water (true or false)}
\item{other_non_dw_source_use}{respondent uses at least one source besides primary non-drinking water source (true or false)}
\item{other_non_dw_sources_packaged}{respondent uses packaged water as an additional source of non-drinking water (true or false)}
\item{other_non_dw_sources_piped_to_home}{respondent uses water piped to home as an additional source of non-drinking water (true or false)}
\item{other_non_dw_sources_piped_to_neighbor}{respondent uses water piped to a neighbor's home as an additional source of non-drinking water (true or false)}
\item{other_non_dw_sources_commercial_tap}{respondent uses a commercial or public tap as an additional source of non-drinking water (true or false)}
\item{other_non_dw_sources_piped_to_compound}{respondent uses water piped to the compound as an additional source of non-drinking water (true or false)}
\item{other_non_dw_sources_borehole}{respondent uses a borehole as an additional source of non-drinking water (true or false)}
\item{other_non_dw_sources_dug_well}{respondent uses a dug well as an additional source of non-drinking water (true or false)}
\item{other_non_dw_sources_spring_water}{respondent uses spring water as an additional source of non-drinking water (true or false)}
\item{other_non_dw_sources_delivered_water}{respondent uses delivered water as an additional source of non-drinking water (true or false)}
\item{other_non_dw_sources_other}{respondent uses another source as an additional source of non-drinking water (true or false)}
\item{secondary_source_reason_availability}{respondent uses a secondary source of non-drinking water because the primary source is not available (true or false)}
\item{secondary_source_reason_unclean}{respondent uses a secondary source of non-drinking water because the primary source is not clean (true or false)}
\item{secondary_source_reason_crowding}{respondent uses a secondary source of non-drinking water because the primary source is crowded (true or false)}
\item{secondary_source_reason_shower}{respondent uses a secondary source of non-drinking water because shower stalls are available (true or false)}
\item{secondary_source_reason_convenient_location}{respondent uses a secondary source of non-drinking water because of its convenient location (true or false)}
\item{tap_payment_mode}{respondent's mechanism for paying for piped water (all respondents use piped water as a primary or secondary source). Options include \code{1} pay_to_fetch: paying to fetch, \code{2} shares_bill: sharing or paying the whole bill, or \code{3} both (at different taps).}
\item{daily_hh_water_cost_for_pay_to_fetch}{daily estimated cost of drinking water for respondent's household}
\item{daily_hh_water_cost_phhm_for_pay_to_fetch}{daily estimated cost of drinking water for respondent's household per household member}
\item{past_struggle_to_find_water}{respondent has struggled to find water before (defined as extreme difficulty in accessing water) (true or false)}
\item{time_of_last_struggle_to_find_water}{respondent's last time of struggle to find water, options including \code{1} last_3_days, \code{2} last_7_days, \code{3} last_30_days, \code{4} last_year, and \code{5} over_year_ago.}
\item{weekdays_struggle_to_find_water}{days in a week the respondent typically struggles to find or pay for water}
\item{past_struggle_primary_reason}{primary reason for past struggles to find water, options including \code{1} availability: availability, \code{2} cost, and \code{3} distance: distance to nearest source.}
\item{tap_closure_knowledge_known}{respondent's knowledge about tap closures: usually known (true or false)}
\item{tap_closure_knowledge_sometimes}{respondent's knowledge about tap closures: sometimes known (true or false)}
\item{tap_closure_knowledge_expected_pattern}{respondent's knowledge about tap closures: expected due to patterns in closures (true or false)}
\item{tap_closure_knowledge_unknown}{respondent's knowledge about tap closures: not known (true or false)}
\item{tap_closure_knowledge_no_answer}{respondent's knowledge about tap closures: no answer (true or false)}
\item{coping_mechanism_spend_more}{respondent copes with water shortage by spending more on the same amount of water (true or false)}
\item{coping_mechanism_purchase_more_to_store_at_home}{respondent copes with water shortage by purchasing extra water to store at home (true or false)}
\item{coping_mechanism_use_other_source}{respondent copes with water shortage by using another source (true or false)}
\item{coping_mechanism_sachet_to_cook}{respondent copes with water shortage by using packaged water for cooking (true or false)}
\item{coping_mechanism_skipped_cooking}{respondent copes with water shortage by skipping cooking (true or false)}
\item{coping_mechanism_sachet_to_bathe}{respondent copes with water shortage by using packaged water for bathing (true or false)}
\item{coping_mechanism_skipped_bathing}{respondent copes with water shortage by skipping bathing (true or false)}
\item{coping_mechanism_closed_business}{respondent copes with water shortage by closing the business (true or false)}
\item{coping_mechanism_skipped_laundry}{respondent copes with water shortage by skipping laundry (true or false)}
\item{water_storage_drinking_water}{respondent typically stores drinking water at home (true or false)}
\item{water_storage_non_drinking_water}{respondent typically stores non-drinking water at home (true or false)}
\item{water_storage_none}{respondent typically does not store water at home (true or false)}
\item{storage_containers_plastic_jug}{respondent stores non-drinking water in plastic jugs (jerry cans or Kufuor gallons) (true or false)}
\item{storage_containers_uncovered_barrels}{respondent stores non-drinking water in uncovered barrels (true or false)}
\item{storage_containers_covered_barrels}{respondent stores non-drinking water in covered barrels (true or false)}
\item{storage_containers_other_covered}{respondent stores non-drinking water in other covered containers (true or false)}
\item{storage_containers_other_uncovered}{respondent stores non-drinking water in other uncovered containers (true or false)}
\item{estimated_non_dw_storage_capacity}{estimated capacity of storage for non-drinking water (liters)}
\item{estimated_stored_non_dw}{estimated actual storage of non-drinking water (liters)}
}
}
\usage{
households
}
\description{
A survey of 116 households was conducted in two low-income communities in metropolitan Accra. These are Korle Gonno, a larger, well-planned coastal area with over 35 household water vendors, and Abuja, a small, densely packed, extralegal settlement with 15 water vendor and bathhouse businesses.
}
\keyword{datasets}
